refactor(ui-components): moved component ui types to correct package
