+++ src/types/ui/index.ts
@@ -10,6 +10,10 @@ export * from './interaction';
 export * from './offline';
 export * from './accessibility';
 export * from './features';
+export * from './alert';
+export * from './switch';
+export * from './radio';
+export * from './checkbox';
 
 // Combined UI State Type for reference
 import type { ToastState } from './toast';
