+++ src/types/ui/switch.ts
