+++ tests/unit/components/ui/Switch.test.tsx
@@ -20,7 +20,7 @@ import {
   AnimatedSwitch,
   SettingsSwitch,
 } from '@/components/ui/switch';
-import type { RecipeSwitchItemProps } from '@/components/ui/switch.types';
+import type { RecipeSwitchItemProps } from '@/types/ui/switch';
 import { Moon, Sun } from 'lucide-react';
 
 describe('Switch Components', () => {
